Update Mario data cloud map position

+++ 05-SceneManager/PlayScene.cpp
@@ -317,7 +317,7 @@ void CPlayScene::SetCam(float cx, float cy)
 		cx = (float)mw - (float)sw;
 
 	//cy -= sh /2 + MARIO_BIG_BBOX_HEIGHT;
-	cy = mh - sh;
+	cy = (float)mh - (float)sh;
 	if (cy <= 0)//Left Edge
 		cy = 0;
